pyproject.toml
[tool.poetry]
name = "stuned"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "stuned", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
jupyter = ">=1.0.0"
absl-py = ">=1.2.0"
accelerate = ">=0.17.1"
aiohttp = ">=3.8.4"
aiosignal = ">=1.3.1"
albumentations = ">=1.3.0"
antlr4-python3-runtime = ">=4.8"
appdirs = ">=1.4.4"
argon2-cffi = ">=21.3.0"
argon2-cffi-bindings = ">=21.2.0"
asttokens = ">=2.0.8"
astunparse = ">=1.6.3"
async-timeout = ">=4.0.2"
attrs = ">=22.1.0"
backcall = ">=0.2.0"
beautifulsoup4 = ">=4.11.1"
bleach = ">=5.0.1"
cachetools = ">=5.2.0"
cffi = ">=1.15.1"
charset-normalizer = ">=2.1.0"
click = ">=8.1.3"
cryptography = ">=39.0.1"
cycler = ">=0.11.0"
debugpy = ">=1.6.3"
decorator = ">=5.1.1"
deepdiff = ">=5.8.1"
defusedxml = ">=0.7.1"
diffusers = ">=0.14.0"
docker-pycreds = ">=0.4.0"
einops = ">=0.3.0"
entrypoints = ">=0.4"
et-xmlfile = ">=1.1.0"
execnet = ">=1.9.0"
executing = ">=1.0.0"
fastjsonschema = ">=2.16.1"
filelock = ">=3.8.0"
flatbuffers = ">=23.1.21"
fonttools = ">=4.36.0"
frozenlist = ">=1.3.3"
fsspec = ">=2023.3.0"
ftfy = ">=6.1.1"
future = ">=0.18.3"
gast = ">=0.4.0"
gdown = ">=4.5.1"
gitdb = ">=4.0.9"
gitpython = ">=3.1.27"
google-api-core = ">=2.11.0"
google-api-python-client = ">=2.79.0"
google-auth = ">=2.16.1"
google-auth-httplib2 = ">=0.1.0"
google-auth-oauthlib = ">=0.4.6"
google-pasta = ">=0.2.0"
googleapis-common-protos = ">=1.58.0"
grpcio = ">=1.51.3"
gspread = ">=5.7.2"
h5pickle = ">=0.4.2"
h5py = ">=3.7.0"
httplib2 = ">=0.21.0"
huggingface-hub = ">=0.13.3"
idna = ">=3.3"
imageio = ">=2.21.1"
importlib-metadata = ">=6.1.0"
iniconfig = ">=1.1.1"
ipykernel = ">=6.15.2"
ipython = ">=8.4.0"
ipython-genutils = ">=0.2.0"
ipywidgets = ">=8.0.1"
jedi = ">=0.18.1"
jinja2 = ">=3.1.2"
joblib = ">=1.1.0"
jsonschema = ">=4.14.0"
jupyter-console = ">=6.4.4"
jupyter-core = ">=4.11.1"
jupyter-client = ">=7.3.5"
jupyterlab-pygments = ">=0.2.2"
jupyterlab-widgets = ">=3.0.2"
#keras = ">=2.11.0"
keras-preprocessing = ">=1.1.2"
kiwisolver = ">=1.4.4"
kornia = ">=0.6.0"
libclang = ">=14.0.6"
lxml = ">=4.9.1"
markdown = ">=3.4.1"
markupsafe = ">=2.1.1"
matplotlib = ">=3.5.3"
matplotlib-inline = ">=0.1.6"
mistune = ">=2.0.4"
multidict = ">=6.0.4"
nbclient = ">=0.6.7"
nbconvert = ">=7.0.0"
nbformat = ">=5.4.0"
nest-asyncio = ">=1.5.5"
networkx = ">=2.8.5"
notebook = ">=6.4.12"
numpy = ">=1.21.1"
nvidia-cublas-cu11 = ">=11.10.3.66"
nvidia-cuda-nvrtc-cu11 = ">=11.7.99"
nvidia-cuda-runtime-cu11 = ">=11.7.99"
nvidia-cudnn-cu11 = ">=8.5.0.96"
oauth2client = ">=4.1.3"
oauthlib = ">=3.2.0"
omegaconf = ">=2.1.1"
open-clip-torch = ">=2.0.2"
opencv-python = ">=4.6.0.66"
opencv-python-headless = ">=4.7.0.72"
openpyxl = ">=3.1.1"
opt-einsum = ">=3.3.0"
ordered-set = ">=4.1.0"
packaging = ">=21.3"
pandas = ">=1.4.3"
pandocfilters = ">=1.5.0"
parso = ">=0.8.3"
pathtools = ">=0.1.2"
pexpect = ">=4.8.0"
pickleshare = ">=0.7.5"
pillow = ">=9.2.0"
pluggy = ">=1.0.0"
prometheus-client = ">=0.14.1"
promise = ">=2.3"
prompt-toolkit = ">=3.0.30"
protobuf = ">=3.19.6"
psutil = ">=5.9.1"
ptyprocess = ">=0.7.0"
pure-eval = ">=0.2.2"
py = ">=1.11.0"
pyasn1 = ">=0.4.8"
pyasn1-modules = ">=0.2.8"
pycparser = ">=2.21"
pydeprecate = ">=0.3.1"
pydrive2 = ">=1.15.1"
pygments = ">=2.13.0"
pyopenssl = ">=23.0.0"
pyparsing = ">=3.0.9"
pyrsistent = ">=0.18.1"
pysocks = ">=1.7.1"
pytest = ">=7.1.3"
pytest-xdist = ">=3.0.2"
python-dateutil = ">=2.8.2"
pytorch-lightning = ">=1.4.2"
pytz = ">=2022.2.1"
pywavelets = ">=1.3.0"
pyyaml = ">=6.0"
pyzmq = ">=23.2.1"
qtconsole = ">=5.3.2"
qtpy = ">=2.2.0"
qudida = ">=0.0.4"
regex = ">=2022.10.31"
requests = ">=2.28.1"
requests-oauthlib = ">=1.3.1"
rsa = ">=4.9"
scikit-image = ">=0.19.3"
scikit-learn = ">=1.1.2"
scipy = ">=1.9.0"
seaborn = ">=0.11.2"
send2trash = ">=1.8.0"
sentry-sdk = ">=1.9.5"
setproctitle = ">=1.3.2"
shortuuid = ">=1.0.9"
six = ">=1.16.0"
sklearn = ">=0.0"
smmap = ">=5.0.0"
soupsieve = ">=2.3.2.post1"
stable-diffusion-sdkit = ">=2.1.4"
stack-data = ">=0.5.0"
tensorboard = ">=2.11.2"
tensorboard-data-server = ">=0.6.1"
tensorboard-plugin-wit = ">=1.8.1"
#tensorflow = ">=2.11.0"
tensorflow-estimator = ">=2.11.0"
tensorflow-io = ">=0.31.0"
tensorflow-io-gcs-filesystem = ">=0.31.0"
termcolor = ">=2.0.1"
terminado = ">=0.15.0"
test-tube = ">=0.7.5"
threadpoolctl = ">=3.1.0"
tifffile = ">=2022.8.12"
timm = ">=0.6.7"
tinycss2 = ">=1.1.1"
tokenizers = ">=0.13.2"
tomli = ">=2.0.1"
torch = ">=1.13.1"
torchmetrics = ">=0.6.0"
torchvision = ">=0.14.1"
tornado = ">=6.2"
tqdm = ">=4.64.0"
traitlets = ">=5.3.0"
transformers = ">=4.26.1"
typing-extensions = ">=4.3.0"
uritemplate = ">=4.1.1"
urllib3 = ">=1.26.11"
wandb = ">=0.13.10"
watchdog = ">=3.0.0"
wcwidth = ">=0.2.5"
webencodings = ">=0.5.1"
werkzeug = ">=2.2.2"
widgetsnbextension = ">=4.0.2"
wrapt = ">=1.11.0"
yarl = ">=1.8.2"
zipp = ">=3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
